Extract the tasks join from /submissions into getSubmissionsWithTasks.js
Move the tasks query and per-submission join into their own module, leaving
the route as a thin adapter, so the logic can be tested directly against a
real Parse Server instead of only through HTTP. The characterization tests
from the previous commit stay green through the move.
